fixes for Int64

+++ src/eventing.cr
@@ -1,9 +1,9 @@
-alias Event = Tuple(Symbol, Int128)
+alias Event = Tuple(Symbol, Int64)
 
 class Eventing
   def initialize
-    @serial = Atomic(Int128).new 0
-    @channels = Hash(Int128, Channel(Event)).new
+    @serial = Atomic(Int64).new 0
+    @channels = Hash(Int64, Channel(Event)).new
   end
 
   def register_channel (channel)
